v1/requirements.txt
matplotlib==3.4.2
numpy==1.21.2
Pillow==9.0.1
pydicom==2.2.2
scikit_image==0.18.3
scipy==1.7.1
SimpleITK==2.1.1
skimage==0.0
tensorflow==2.4.1
tqdm==4.62.3
